requirements.txt
Flask==3.0.0
google-cloud-speech==2.21.0
gunicorn==21.2.0